+++ lib/src/special_inline_span_base.dart
@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 import 'package:flutter/rendering.dart';
 
 ///